promptxecure-core/pyproject.toml
[project]
name = "promptxecure-core"
version = "0.1.0"
description = "Multi-layered prompt injection detection engine for LLMs"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "PromptXecure Team"}]

dependencies = [
    "pyyaml>=6.0",
    "sentence-transformers>=3.0.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "litellm>=1.30.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "bandit>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/promptxecure_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
